+++ drivers/ESP32/bdring_i2s_6_axis_map.h
@@ -107,8 +107,6 @@
 // Define probe switch input pin.
 #if PROBE_ENABLE
 #define PROBE_PIN       GPIO_NUM_25
-#else
-#define PROBE_PIN       0xFF
 #endif
 
 #if KEYPAD_ENABLE