+++ drivers/ESP32/bdring_v3.5_map.h
@@ -84,8 +84,6 @@
 // Define probe switch input pin.
 #if PROBE_ENABLE
 #define PROBE_PIN       GPIO_NUM_32
-#else
-#define PROBE_PIN       0xFF
 #endif
 
 #if KEYPAD_ENABLE