+++ drivers/ESP32/bdring_v4_map.h
@@ -91,8 +91,6 @@
 // Define probe switch input pin.
 #if PROBE_ENABLE
 #define PROBE_PIN       GPIO_NUM_32
-#else
-#define PROBE_PIN       0xFF
 #endif
 
 #if MODBUS_ENABLE