+++ drivers/STM32F4xx/Src/i2c.c
@@ -62,6 +62,9 @@ void i2c_init (void)
     __HAL_RCC_I2C1_CLK_ENABLE();
 
     HAL_I2C_Init(&i2c_port);
+
+    HAL_NVIC_EnableIRQ(I2C1_EV_IRQn);
+    HAL_NVIC_EnableIRQ(I2C1_ER_IRQn);
 #endif
 
 #if I2C_PORT == 2
@@ -75,8 +78,33 @@ void i2c_init (void)
     __HAL_RCC_I2C2_CLK_ENABLE();
 
     HAL_I2C_Init(&i2c_port);
+
+    HAL_NVIC_EnableIRQ(I2C2_EV_IRQn);
+    HAL_NVIC_EnableIRQ(I2C2_ER_IRQn);
 #endif
 }
+
+#if I2C_PORT == 1
+void I2C1_EV_IRQHandler(void)
+{
+  HAL_I2C_EV_IRQHandler(&i2c_port);
+}
+
+void I2C1_ER_IRQHandler(void)
+{
+  HAL_I2C_ER_IRQHandler(&i2c_port);
+}
+#else
+void I2C2_EV_IRQHandler(void)
+{
+  HAL_I2C_EV_IRQHandler(&i2c_port);
+}
+
+void I2C2_ER_IRQHandler(void)
+{
+  HAL_I2C_ER_IRQHandler(&i2c_port);
+}
+#endif
 
 #endif
 
@@ -113,10 +141,7 @@ void I2C_GetKeycode (uint32_t i2cAddr, keycode_callback_ptr callback)
     keycode = 0;
     keypad_callback = callback;
 
-    HAL_StatusTypeDef ret = HAL_I2C_Master_Receive_IT(&i2c_port, KEYPAD_I2CADDR << 1, &keycode, 1);
-
-    if(!ret)
-        ret = HAL_I2C_Master_Receive_IT(&i2c_port, KEYPAD_I2CADDR << 1, &keycode, 1);
+    HAL_I2C_Master_Receive_IT(&i2c_port, KEYPAD_I2CADDR << 1, &keycode, 1);
 }
 
 void HAL_I2C_MasterRxCpltCallback(I2C_HandleTypeDef *hi2c)